move fetching results to the a separate service
